Promote compute_tangential_wind to common (3-way dedup)
The same RBF edge-tangential-wind stencil existed in tmx (wp), dycore
(vp cast) and advection (compute_edge_tangential, with a TODO noting
the duplication). Promote the better-documented tmx version to
common/interpolation/stencils with both wp and vp variants under their
existing public names, and delete all three originals.
