+++ model/atmosphere/dycore/src/icon4py/model/atmosphere/dycore/stencils/compute_avg_vn_and_graddiv_vn_and_vt.py
@@ -8,14 +8,14 @@
 import gt4py.next as gtx
 from gt4py.next import astype, neighbor_sum
 
-from icon4py.model.atmosphere.dycore.stencils.compute_tangential_wind import (
-    _compute_tangential_wind,
-)
 from icon4py.model.atmosphere.dycore.stencils.spatially_average_flux_or_velocity import (
     _spatially_average_flux_or_velocity,
 )
 from icon4py.model.common import dimension as dims, field_type_aliases as fa
 from icon4py.model.common.dimension import E2C2EO, E2C2EODim
+from icon4py.model.common.interpolation.stencils.compute_tangential_wind import (
+    _compute_tangential_wind,
+)
 from icon4py.model.common.type_alias import vpfloat, wpfloat
